+++ include/qpdf/QPDFObjectHandle.hh
@@ -334,6 +334,15 @@ class QPDFObjectHandle
     QPDF_DLL
     inline bool isInitialized() const;
 
+    // Two QPDFObjectHandle objects are equal if they point to exactly
+    // the same underlying object, meaning that changes to one are
+    // reflected in the other, or "if you paint one, the other one
+    // changes color."
+    QPDF_DLL
+    bool operator==(QPDFObjectHandle const&) const;
+    QPDF_DLL
+    bool operator!=(QPDFObjectHandle const&) const;
+
     // Return type code and type name of underlying object.  These are
     // useful for doing rapid type tests (like switch statements) or
     // for testing and debugging.